Create Longest_Common_Prefix.py
This function takes an array of strings as an argument and returns the longest common prefix amongst the words. If there is no common prefix, an empty string is returned. Prefix here is intended to mean a common letter or letters across all words present in the array, in the correct order, at the beginning of the word. This doesn't necessarily mean a prefix in the etymological sense.