Create Valid_Parentheses.py
This code takes in a string argument containing parentheses ( ), brackets [ ], and braces { }. The string is considered valid if all opening brackets are closed, and if all brackets are closed in the corresponding order to how they were opened. The function checks to see if these parameters are met and returns True if they are and False if they are not.