Create Remove_Duplicates.py
This function takes an ordered array argument. The function reorders the array so there is only one instance of each number, and those numbers are kept in the same relative order. All repeat numbers are changed to an underscore and moved to the back of the array so that the first k values of the array are the k unique numbers present in the array. The number of unique values is returned.